Add debug output to fish-persistent-data onCreate
fish-persistent-data feature fails to execute the onCreate.sh script
when the feature is enabled as user defaultFeatures and a recovery
container is created. This is a special environment and cumbersome to
debug, but this version hopefully prints out enough information to
provide a clue at the exact problem and how it should be handled
without failure.